pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Phase 2 Todo App Backend API - FastAPI with SQLModel and Neon PostgreSQL"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.24",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "pyjwt>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "pytest>=9.0.2",
    "httpx>=0.28.1",
    "openai-agents>=0.6.3",
    "mcp>=1.24.0",
    "litellm>=1.80.10",
    "sse-starlette>=2.2.0",
    "alembic>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",  # For TestClient
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
